Group RandomField, BigInt, Int, Uint, FieldConfig and ConfigRef into single module
Refactor conversion module: reorganize and implement From traits for RandomField in field module

Rename Config::I to Config::B for constistency
