--- src/crypto_int/mod.rs
@@ -1,5 +1,0 @@
-mod int;
-mod uint;
-
-pub use int::Int;
-pub use uint::Uint;
