+++ src/field/biginteger/arithmetic.rs
@@ -119,8 +119,9 @@ pub fn mac_discard(a: u64, b: u64, c: u64, carry: &mut u64) {
 
 macro_rules! mac_with_carry {
     ($a:expr, $b:expr, $c:expr, &mut $carry:expr$(,)?) => {{
-        let tmp =
-            ($a as u128) + $crate::biginteger::arithmetic::widening_mul($b, $c) + ($carry as u128);
+        let tmp = ($a as u128)
+            + $crate::field::biginteger::arithmetic::widening_mul($b, $c)
+            + ($carry as u128);
         $carry = (tmp >> 64) as u64;
         tmp as u64
     }};
@@ -128,7 +129,7 @@ macro_rules! mac_with_carry {
 
 macro_rules! mac {
     ($a:expr, $b:expr, $c:expr, &mut $carry:expr$(,)?) => {{
-        let tmp = ($a as u128) + $crate::biginteger::arithmetic::widening_mul($b, $c);
+        let tmp = ($a as u128) + $crate::field::biginteger::arithmetic::widening_mul($b, $c);
         $carry = (tmp >> 64) as u64;
         tmp as u64
     }};
