+++ src/lib.rs
@@ -1,11 +1,9 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
-pub mod biginteger;
 pub mod ccs;
 mod const_helpers;
-pub mod crypto_int;
+pub mod conversion;
 pub mod field;
-pub mod field_config;
 pub mod poly;
 pub mod poly_f;
 pub mod poly_z;
